+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Changelog
 =========
 
-## 1.1.*
+## 1.1.24 (2020-06-06)
+ * Added support for Robots.txt module
  * Fontface in dashboard
  * Fixed: Dropdown menu hidding by popover is show
  * Check height of dropdown, dropup variation